requirements.txt
python-dotenv

flask
pymongo